chore: Properly route to root element on iOS and WASM
